+++ src/main/java/com/example/tily/step/StepRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface StepRepository extends JpaRepository<Step, Integer> {
+public interface StepRepository extends JpaRepository<Step, Long> {
     List<Step> findByRoadmap(Roadmap roadmap); // 나중에 사용
 }
